Integrate verification helper into parser
