+++ src/xcode/ENA/ENA/Source/Scenes/QRScanner/_tests_/CheckinQRCodeParserTests.swift
@@ -6,10 +6,9 @@ import XCTest
 @testable import ENA
 
 // IMPORTANT: THESE TESTS ARE BASED ON THE CURRENT EXPECTED REGEX, WE NEED TO UPDATE THEM IF THE REGEX IS UPDATED
-class QRCodeVerificationHelperTests: CWATestCase {
+class CheckinQRCodeParserTests: CWATestCase {
 
     func testValidURL() {
-		let appConfig = CachedAppConfigurationMock()
 		let validHostName = "https://e.coronawarn.app"
 		let validVersion = "?v=1"
 		let validPayload = "#CAESRggBEi1CdXJsaW5ndG9uIENvYXQgRmFjdG9yeSBXYXJlaG91c2UgQ29ycG9yYXRpb24aDzE5OTAgVGlrb2QgUGlrZSgAMAAacQgBElswWTATBgcqhkjOPQIBBggqhkjOPQMBBwNCAARzsMSy1xQhFzKTfk5gMn3n+FODdRWGxoNcpPTMrs2Ec9ejLOKkSc6ncbI1cNWlo+LdwB9CbF64UxkAhfb7oDleGhAgd3QGMXeySb/sMUUkwZMgIgcIARAHGNsK"
@@ -17,10 +16,11 @@ class QRCodeVerificationHelperTests: CWATestCase {
 
 		let onSuccessExpectation = expectation(description: "onSuccess called")
 
-		let qrCodeVerificationHelper = QRCodeVerificationHelper()
-		qrCodeVerificationHelper.verifyQrCode(
+		let checkinQRCodeParser = CheckinQRCodeParser(
+			appConfigurationProvider: CachedAppConfigurationMock()
+		)
+		checkinQRCodeParser.verifyQrCode(
 			qrCodeString: validURL,
-			appConfigurationProvider: appConfig,
 			onSuccess: { _ in
 				onSuccessExpectation.fulfill()
 			},
@@ -30,17 +30,17 @@ class QRCodeVerificationHelperTests: CWATestCase {
 	}
 	
 	func testInvalidURL_noVersion() {
-		let appConfig = CachedAppConfigurationMock()
 		let validHostName = "https://e.coronawarn.app"
 		let validPayload = "#CAESJQgBEgpBZ3dheSBJbmMuGhExNTk0IERlZmZlIEF2ZW51ZSgAMAAadggBEmA4xNrp5hKJoO_yVbXfF1gS8Yc5nURhOIVLG3nUcSg8IPsI2e8JSIhg-FrHUymQ3RR80KUKb1lZjLQkfTUINUP16r6-jFDURwUlCQQi6NXCgI0rQw0a4MrVrKMbF4NzhQMaEPXDJZ2XSeO0SY43-KCQlQciBggBEAQYHA"
 		let validURL = validHostName + validPayload
 
 		let onErrorExpectation = expectation(description: "onError called")
 
-		let qrCodeVerificationHelper = QRCodeVerificationHelper()
-		qrCodeVerificationHelper.verifyQrCode(
+		let checkinQRCodeParser = CheckinQRCodeParser(
+			appConfigurationProvider: CachedAppConfigurationMock()
+		)
+		checkinQRCodeParser.verifyQrCode(
 			qrCodeString: validURL,
-			appConfigurationProvider: appConfig,
 			onSuccess: { _ in },
 			onError: { error in
 				XCTAssertEqual(error, .codeNotFound, "Invalid url code should be: .codeNotFound")
@@ -59,10 +59,11 @@ class QRCodeVerificationHelperTests: CWATestCase {
 
 		let onErrorExpectation = expectation(description: "onError called")
 
-		let qrCodeVerificationHelper = QRCodeVerificationHelper()
-		qrCodeVerificationHelper.verifyQrCode(
+		let checkinQRCodeParser = CheckinQRCodeParser(
+			appConfigurationProvider: CachedAppConfigurationMock()
+		)
+		checkinQRCodeParser.verifyQrCode(
 			qrCodeString: validURL,
-			appConfigurationProvider: appConfig,
 			onSuccess: { _ in },
 			onError: { error in
 				XCTAssertEqual(error, .codeNotFound, "Invalid url code should be: .codeNotFound")
@@ -73,17 +74,17 @@ class QRCodeVerificationHelperTests: CWATestCase {
 	}
 	
 	func testInvalidURL_WrongPayload() {
-		let appConfig = CachedAppConfigurationMock()
 		let validHostName = "https://e.coronawarn.app"
 		let validVersion = "?v=1"
 		let validPayload = "#Wrong_payload"
 		let validURL = validHostName + validVersion + validPayload
 		let onErrorExpectation = expectation(description: "onError called")
 
-		let qrCodeVerificationHelper = QRCodeVerificationHelper()
-		qrCodeVerificationHelper.verifyQrCode(
+		let checkinQRCodeParser = CheckinQRCodeParser(
+			appConfigurationProvider: CachedAppConfigurationMock()
+		)
+		checkinQRCodeParser.verifyQrCode(
 			qrCodeString: validURL,
-			appConfigurationProvider: appConfig,
 			onSuccess: { _ in },
 			onError: { error in
 				XCTAssertEqual(error, .invalidVendorData, "Invalid url code should be: .invalidVendorData")
@@ -102,8 +103,10 @@ class QRCodeVerificationHelperTests: CWATestCase {
 			)
 		let onSuccessExpectation = expectation(description: "onSuccess called")
 
-		let qrCodeVerificationHelper = QRCodeVerificationHelper()
-		qrCodeVerificationHelper.validateTraceLocationInformation(
+		let checkinQRCodeParser = CheckinQRCodeParser(
+			appConfigurationProvider: CachedAppConfigurationMock()
+		)
+		checkinQRCodeParser.validateTraceLocationInformation(
 			traceLocation: traceLocation,
 			onSuccess: { _ in
 				onSuccessExpectation.fulfill()
@@ -121,8 +124,10 @@ class QRCodeVerificationHelperTests: CWATestCase {
 			)
 		let onErrorExpectation = expectation(description: "onError called")
 
-		let qrCodeVerificationHelper = QRCodeVerificationHelper()
-		qrCodeVerificationHelper.validateTraceLocationInformation(
+		let checkinQRCodeParser = CheckinQRCodeParser(
+			appConfigurationProvider: CachedAppConfigurationMock()
+		)
+		checkinQRCodeParser.validateTraceLocationInformation(
 			traceLocation: traceLocation,
 			onSuccess: { _ in },
 			onError: { error in
@@ -141,8 +146,10 @@ class QRCodeVerificationHelperTests: CWATestCase {
 			)
 		let onErrorExpectation = expectation(description: "onError called")
 
-		let qrCodeVerificationHelper = QRCodeVerificationHelper()
-		qrCodeVerificationHelper.validateTraceLocationInformation(
+		let checkinQRCodeParser = CheckinQRCodeParser(
+			appConfigurationProvider: CachedAppConfigurationMock()
+		)
+		checkinQRCodeParser.validateTraceLocationInformation(
 			traceLocation: traceLocation,
 			onSuccess: { _ in },
 			onError: { error in
@@ -161,8 +168,10 @@ class QRCodeVerificationHelperTests: CWATestCase {
 			)
 		let onErrorExpectation = expectation(description: "onError called")
 
-		let qrCodeVerificationHelper = QRCodeVerificationHelper()
-		qrCodeVerificationHelper.validateTraceLocationInformation(
+		let checkinQRCodeParser = CheckinQRCodeParser(
+			appConfigurationProvider: CachedAppConfigurationMock()
+		)
+		checkinQRCodeParser.validateTraceLocationInformation(
 			traceLocation: traceLocation,
 			onSuccess: { _ in },
 			onError: { error in
@@ -181,8 +190,10 @@ class QRCodeVerificationHelperTests: CWATestCase {
 			)
 		let onErrorExpectation = expectation(description: "onError called")
 
-		let qrCodeVerificationHelper = QRCodeVerificationHelper()
-		qrCodeVerificationHelper.validateTraceLocationInformation(
+		let checkinQRCodeParser = CheckinQRCodeParser(
+			appConfigurationProvider: CachedAppConfigurationMock()
+		)
+		checkinQRCodeParser.validateTraceLocationInformation(
 			traceLocation: traceLocation,
 			onSuccess: { _ in },
 			onError: { error in
@@ -201,8 +212,10 @@ class QRCodeVerificationHelperTests: CWATestCase {
 			)
 		let onErrorExpectation = expectation(description: "onError called")
 
-		let qrCodeVerificationHelper = QRCodeVerificationHelper()
-		qrCodeVerificationHelper.validateTraceLocationInformation(
+		let checkinQRCodeParser = CheckinQRCodeParser(
+			appConfigurationProvider: CachedAppConfigurationMock()
+		)
+		checkinQRCodeParser.validateTraceLocationInformation(
 			traceLocation: traceLocation,
 			onSuccess: { _ in },
 			onError: { error in
@@ -221,8 +234,10 @@ class QRCodeVerificationHelperTests: CWATestCase {
 			)
 		let onErrorExpectation = expectation(description: "onError called")
 
-		let qrCodeVerificationHelper = QRCodeVerificationHelper()
-		qrCodeVerificationHelper.validateTraceLocationInformation(
+		let checkinQRCodeParser = CheckinQRCodeParser(
+			appConfigurationProvider: CachedAppConfigurationMock()
+		)
+		checkinQRCodeParser.validateTraceLocationInformation(
 			traceLocation: traceLocation,
 			onSuccess: { _ in },
 			onError: { error in
@@ -241,8 +256,10 @@ class QRCodeVerificationHelperTests: CWATestCase {
 			)
 		let onErrorExpectation = expectation(description: "onError called")
 
-		let qrCodeVerificationHelper = QRCodeVerificationHelper()
-		qrCodeVerificationHelper.validateTraceLocationInformation(
+		let checkinQRCodeParser = CheckinQRCodeParser(
+			appConfigurationProvider: CachedAppConfigurationMock()
+		)
+		checkinQRCodeParser.validateTraceLocationInformation(
 			traceLocation: traceLocation,
 			onSuccess: { _ in },
 			onError: { error in
@@ -265,8 +282,10 @@ class QRCodeVerificationHelperTests: CWATestCase {
 		)
 		let onErrorExpectation = expectation(description: "onError called")
 
-		let qrCodeVerificationHelper = QRCodeVerificationHelper()
-		qrCodeVerificationHelper.validateTraceLocationInformation(
+		let checkinQRCodeParser = CheckinQRCodeParser(
+			appConfigurationProvider: CachedAppConfigurationMock()
+		)
+		checkinQRCodeParser.validateTraceLocationInformation(
 			traceLocation: traceLocation,
 			onSuccess: { _ in },
 			onError: { error in
